pyproject.toml
[project]
name = "deepnotes"
version = "0.1.0"
dependencies = [
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "pandas>=2.0"
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.ruff]
lint.select = ["F", "E", "W", "I", "B", "C4", "PL"]
lint.ignore = [
  "B904",  # abstract base class with empty __slots__
  "E501",  # line-too-long (handled by formatter)
  "PLR0913"  # too-many-arguments
]
line-length = 88
exclude = [
  ".venv",
  "__pycache__",
  "build",
  "dist"
]
target-version = "py312"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = "dynamic" 